put the schema builder traits & utilities in a separate crate
