+++ serde_avro_derive_macros/Cargo.toml
@@ -3,16 +3,20 @@
 	description = "Derive avro schema for Rust structs for serde_avro_fast"
 	edition = "2021"
 	license = "LGPL-3.0-only"
-	name = "serde_avro_derive"
+	name = "serde_avro_derive_macros"
 	repository = "https://github.com/Ten0/serde_avro_fast"
 	version = "0.1.0"
 	workspace = ".."
 
+[lib]
+	proc-macro = true
+
 [dependencies]
-	serde_avro_derive_macros = { path = "../serde_avro_derive_macros" }
-	serde_avro_fast = { path = "../serde_avro_fast" }
+	darling = "0.20"
+	proc-macro2 = "1"
+	quote = "1"
+	syn = { version = "2", features = ["visit-mut"] }
 
 [dev-dependencies]
-	pretty_assertions = "1"
 	serde_avro_derive = { path = "../serde_avro_derive" }
 	serde_json = "1"
